So I was thinking, why do I have a form, when I don't use a form? The logic probably fits better under the model anyway. So I dropped my TicBoardForm, and instead just use the TicBoard model. I still need to clean up some of the methods I have moved from the form to the model.
