+++ asite/tic/views.py
@@ -1,5 +1,4 @@
 from django.shortcuts import render, get_object_or_404
-from tic.forms import TicBoardForm
 from tic.models import TicBoard
 from django.forms.models import model_to_dict
 
@@ -7,36 +6,38 @@ from django.forms.models import model_to_dict
 
 
 def index(request):
-
+	
 	return render(request,'tic/index.html',{})	
 
-def processBoard(request,opick,ticboard):
-	mtd=model_to_dict(ticboard)
-	if opick:		
+def processBoard(request,opick,board):
+	
+	mtd=model_to_dict(board)
+	if opick:	
+		board.opick='o'
+		board.save()	
 		mtd[opick]='o'
-
-	form=TicBoardForm()
-	xpick,winners=form.mkPick(mtd)
+		
+	xpick,winners=board.mkPick(mtd)
 	mtd[xpick]='x'
-	form =TicBoardForm(mtd, instance=ticboard)			
-	if form.is_valid():
-		board=form.save()	
-		board.save()		
-	myargs={'pk':board.pk,'wingroup':winners,'xpick':xpick,'form':form}	
+	board.xpick='x'
+	board.save()		
+	myargs={'pk':board.pk,'wingroup':winners,'xpick':xpick}	
 	return myargs
 	
 
 
 
 def createBoard(request):
-	ticboard=TicBoard()
-	ticboard.save()
-	myargs=processBoard(request,False,ticboard)
+	
+	board=TicBoard()
+	board.save()
+	myargs=processBoard(request,False,board)
 	return render(request,'tic/tic.js',myargs)	
 
 
 
 def updateBoard(request,pk,opick):
-	ticboard=get_object_or_404(TicBoard, pk=pk)
-	myargs=processBoard(request,opick,ticboard)
+	
+	board=get_object_or_404(TicBoard, pk=pk)
+	myargs=processBoard(request,opick,board)
 	return render(request,'tic/tic.js',myargs)
